+++ src/js/pyodide.js
@@ -10,6 +10,7 @@ import {
 } from "./load-pyodide.js";
 import { makePublicAPI, registerJsModule } from "./api.js";
 import "./pyproxy.gen.js";
+import "./error_handling.gen.js";
 
 /**
  * @typedef {import('./pyproxy.gen').PyProxy} PyProxy